feat(sidebar): separate internal and custom nav groups
Move custom sidebar links out of internal navigation and render them as a separate MORE group in dashboard/server routers.
