+++ resources/scripts/routers/DashboardRouter.tsx
@@ -9,7 +9,7 @@ import routes from '@/routers/routes';
 import { RouterContainer } from '@/reviactyl/ui/RouterContainer';
 import Navbar from '@/reviactyl/ui/Navbar';
 import { LogoContainer } from '@/reviactyl/ui/LogoContainer';
-import { XIcon, MenuIcon } from '@heroicons/react/solid';
+import { XIcon, MenuIcon, ExternalLinkIcon } from '@heroicons/react/solid';
 import tw from 'twin.macro';
 import { ContentContainer } from '@/reviactyl/ui/ContentContainer';
 import { motion } from 'framer-motion';
@@ -21,6 +21,8 @@ import MaintenanceAlert from '@/reviactyl/ui/MaintenanceAlert';
 import QuickLinks from '@/reviactyl/ui/QuickLinks';
 import Maintenance from '@/reviactyl/ui/Maintenance';
 import { useTranslation } from 'react-i18next';
+import { FaHouse } from 'react-icons/fa6';
+import { ReviactylSidebarButton } from '@/state/reviactyl';
 
 interface Props {
     route: any;
@@ -42,13 +44,50 @@ const NavItem = ({ route }: Props) => {
 };
 
 const DashboardNavigation = () => {
+    const { t } = useTranslation('routes');
+    const customSidebarButtons = useStoreState((state) => state.reviactyl.data?.sidebarButtons ?? []);
+    const normalizedSidebarButtons = (Array.isArray(customSidebarButtons) ? customSidebarButtons : []).filter(
+        (button): button is ReviactylSidebarButton =>
+            typeof button?.label === 'string' &&
+            button.label.trim().length > 0 &&
+            typeof button?.url === 'string' &&
+            button.url.trim().length > 0
+    );
+
     return (
         <>
-            {routes.account
-                .filter((route) => !!route.name)
-                .map((route) => (
-                    <NavItem key={route.name} route={route} />
-                ))}
+            <div>
+                <Navigate id='index.dashboard' to='/' end className='mt-2'>
+                    <span className='flex items-center'>
+                        <FaHouse className='w-5 mr-1' /> {t('index.dashboard')}
+                    </span>
+                </Navigate>
+
+                {routes.account
+                    .filter((route) => !!route.name)
+                    .map((route) => (
+                        <NavItem key={route.name} route={route} />
+                    ))}
+            </div>
+
+            {normalizedSidebarButtons.length > 0 && (
+                <div>
+                    <span className='label'>MORE</span>
+                    {normalizedSidebarButtons.map((button, index) => (
+                        <a
+                            key={`${button.url}-${index}`}
+                            href={button.url}
+                            target={button.newTab === true ? '_blank' : undefined}
+                            rel={button.newTab === true ? 'noopener noreferrer' : undefined}
+                        >
+                            <span className='flex items-center'>
+                                <ExternalLinkIcon className='w-4 h-4 mr-2' />
+                                {button.label}
+                            </span>
+                        </a>
+                    ))}
+                </div>
+            )}
         </>
     );
 };
@@ -95,7 +134,7 @@ function DashboardRouter() {
                             animate={{ opacity: 1 }}
                             transition={{ duration: 0.15, ease: 'easeIn' }}
                         >
-                            <Sidebar isOpen={isSidebarOpen} dashboard>
+                            <Sidebar isOpen={isSidebarOpen}>
                                 <DashboardNavigation />
                             </Sidebar>
                         </motion.div>
